requirements.txt
line-bot-sdk
scikit-image
Flask
```